attached_assets/requirements_1751897065480.txt
# Core dependencies
python-telegram-bot==20.7
asyncio
python-dotenv==1.0.0

# Cryptography
cryptography==41.0.7

# Networking
pyzmq==25.1.2

# Data handling
pandas==2.1.4
numpy==1.26.2

# Optional: for better logging
colorlog==6.8.0